refactor(cli): split commands/mod.rs into per-domain modules
mod.rs (2294 lines) becomes a thin re-export hub. Command bodies move to
index/search/graph/rag/manage modules; cross-cutting helpers to shared.rs
and progress.rs; shared test fixtures to test_support.rs. main.rs paths
preserved via re-exports; all tests moved intact (no behavior change).

Also gate `use super::remote` in doctor.rs behind the remote-s3 feature:
the prior glob import never warned, but the explicit import is unused
without the feature.
